fix type inconsistency

+++ compiler/base/actions.cpp
@@ -142,7 +142,7 @@ namespace Base {
         // The help menu!
         void printHelpMenu() {
             REPORT(Console::START_REPORT, Console::NORMAL_REPORT, "Options:\n");
-            for (int i = 0; i < actions.size(); i++) {
+            for (size_t i = 0; i < actions.size(); i++) {
                 REPORT("    ", actions[i].info[0], "/", actions[i].info[1], " \t ", actions[i].info[2]);
                 for (int j = 0; j < actions[i].input.size(); j++) {
                     REPORT("\n    \t", actions[i].input[j]);
